requirements.txt
cec==0.2.7
dbus-python==1.2.16